server/requirements.txt
Django==1.11.4
djangorestframework==3.5.1